% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanonymize.r
\name{connectToPostgres}
\alias{connectToPostgres}
\title{ConnectToPostgres}
\usage{
connectToPostgres(connstring)
}
\description{
Local helper function to attach to Postgres. Must have run init.sh previously
}
\examples{
connectToPostgres(connstring)
}
