% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanonymize.r
\name{getConnectionString}
\alias{getConnectionString}
\title{GetConnectionString}
\usage{
getConnectionString()
}
\description{
Retrieves the connection string either from env or dbfs
}
\examples{
getConnectionString()
}
