% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanonymize.r
\name{selectNextTable}
\alias{selectNextTable}
\title{selectNextTable}
\usage{
selectNextTable(desiredState)
}
\description{
Fetches a random table in desiredState
}
\examples{
selectNextTable(desiredState)
}
