% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanonymize.r
\docType{data}
\name{states}
\alias{states}
\title{States}
\format{An object of class \code{environment} of length 4.}
\usage{
states
}
\description{
An Enum that drives the FSM
}
\examples{
states
}
\keyword{datasets}
